# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Blog Go API server.",
        "title": "Blog Go API",
        "contact": {
            "name": "API Support",
            "url": "http://somewhere.com/support",
            "email": "support@somewhere.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/blog": {
            "get": {
                "description": "Get All Blog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Get All Blog",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blog Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.jsonResponseWithPaginate"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Blog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Update Blog",
                "parameters": [
                    {
                        "description": "Update information",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateBlogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.jsonResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Blog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Create Blog",
                "parameters": [
                    {
                        "description": "Create information",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateBlogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.jsonResponse"
                        }
                    }
                }
            }
        },
        "/api/blog/id": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Blog By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Get Blog By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Blog ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.jsonResponse"
                        }
                    }
                }
            }
        },
        "/api/blog/tag": {
            "get": {
                "description": "Get All Blog With Tag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Get All Blog With Tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blog Title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tag Name",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.jsonResponseWithPaginate"
                        }
                    }
                }
            }
        },
        "/api/blog/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Blog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Delete Blog",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Blog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.jsonResponse"
                        }
                    }
                }
            }
        },
        "/api/blog/{url}": {
            "get": {
                "description": "Get Blog By URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blog"
                ],
                "summary": "Get Blog By URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blog URL",
                        "name": "url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.jsonResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.BlogTagRequest": {
            "type": "object",
            "required": [
                "blog_id",
                "id",
                "tag_id"
            ],
            "properties": {
                "blog_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "deleted": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "tag_id": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "api.CreateBlogRequest": {
            "type": "object",
            "required": [
                "content",
                "image",
                "title",
                "url"
            ],
            "properties": {
                "blog_tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.BlogTagRequest"
                    }
                },
                "content": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "api.UpdateBlogRequest": {
            "type": "object",
            "required": [
                "content",
                "id",
                "image",
                "title",
                "url"
            ],
            "properties": {
                "blog_tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.BlogTagRequest"
                    }
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "image": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "api.jsonResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.jsonResponseWithPaginate": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    }
}